Fix compiler warnings, type safety, serialization, and modernize codebase patterns

+++ src/main/java/ch/technokrat/gecko/geckocircuits/control/ControlLimit.java
@@ -126,7 +126,7 @@ public final class ControlLimit extends ControlBlock implements ControlInputTwoT
             graphics.drawLine(xPos * dpix, (int) ((yPos + 1/2.0) * dpix), xPos * dpix, (yPos + Y_DRAW_OFFSET) * dpix);
             graphics.setColor(GlobalColors.farbeInBearbeitungCONTROL);
             graphics.drawString(MIN, dpix * xPos + X_DRAW_OFFSET, (int) (dpix * (yPos + Y_STRING_DRAW_OFFSET)));
-            graphics.drawString(MAX, (int) (dpix * xPos) + X_DRAW_OFFSET, (int) (dpix * (yPos + Y_STRING_DRAW_OFFSET+1)));
+            graphics.drawString(MAX, dpix * xPos + X_DRAW_OFFSET, (int) (dpix * (yPos + Y_STRING_DRAW_OFFSET+1)));
             graphics.setColor(origColor);
         }
     }
